Beta 19, fixed XML errors, run clean-up tool, started translating backstories
After update to Beta 19, a lot of XML erros came up. These are fixed.
A lot of translations seemed bad (mostly missing and switching to english), so the clean-up tool is runned and the report tool showed a lot missing/erros. See readme.md.
A start is made for translating the backstories.